dapp: factor AdGrid into its own component
