+++ src/Homepage.vue
@@ -1,32 +1,4 @@
 <style lang="scss">
-.adGrid {
-  img, .nsfwAd, a div {
-    position: absolute;
-    display: block;
-    overflow: hidden;
-    font-size: 11px;
-    color: rgba(0, 0, 0, 0.7);
-    white-space: nowrap;
-  }
-
-  .nsfwAd {
-    background: #000;
-  }
-
-  &.active img {
-    box-shadow: 0px 0px 0px 1px rgba(255, 255, 255, 0.4) inset;
-  }
-
-  .previewAd {
-    background: rgba(255, 255, 255, 0.8);
-    border: 1px solid rgba(255, 255, 255, 0.9);
-    z-index: 5;
-  }
-  .previewAd.locked {
-    background: #ffcc47;
-  }
-}
-
 .edit {
   display: inline-block;
   margin-top: 1em;
@@ -45,14 +17,7 @@
 
 <template>
   <div class="container">
-    <div :class="{ adGrid: true, active: !!$store.state.previewAd }" :style="gridStyle(prerendered)">
-      <template v-for="ad in $store.state.ads" v-if="ad">
-        <Ad :showNSFW="showNSFW" :ad="ad" :skipImage="!loadRemoteImages"></Ad>
-      </template>
-      <vue-draggable-resizable ref="draggable" :active="true" :minw="10" :minh="10" :x="$store.state.previewAd.x" :y="$store.state.previewAd.y" :w="80" :h="40" :grid="[10,10]" :parent="true" @dragstop="updatePreview" @resizestop="updatePreview" :draggable="!previewLocked" :resizable="!previewLocked" v-if="$store.state.previewAd" v-bind:class="{previewAd: true, locked: previewLocked}">
-        <Buy :web3="web3" :contract="contract" :isReadOnly="isReadOnly" @buy="onBuy"></Buy>
-      </vue-draggable-resizable>
-    </div>
+    <AdGrid :web3="web3" :contract="contract" :showNSFW="showNSFW" :isReadOnly="isReadOnly" :prerendered="prerendered"></AdGrid>
 
     <div class="edit" v-if="$store.state.numOwned > 0">
       {{$store.state.numOwned}} ads owned by you. <button v-on:click="showPublish = true" v-if="!showPublish">Edit Ads</button>
@@ -64,11 +29,9 @@
 </template>
 
 <script>
-import Ad from './Ad.vue'
-import Buy from './Buy.vue'
+import AdGrid from './AdGrid.vue'
 import Publish from './Publish.vue'
 import Offline from './Offline.vue'
-import VueDraggableResizable from 'vue-draggable-resizable'
 
 function toAd(i, r) {
   return {
@@ -90,27 +53,10 @@ export default {
   props: ["web3", "contract", "isReadOnly", "showNSFW", "prerendered"],
   data() {
     return {
-      previewLocked: false,
       showPublish: false,
-      loadRemoteImages: this.prerendered ? this.prerendered.loadRemoteImages : true,
     }
   },
   methods: {
-    onBuy: function (ad) {
-      this.previewLocked = true;
-    },
-    updatePreview(x, y, width, height) {
-      const dragged = this.$refs.draggable;
-      this.$store.commit('updatePreview', {
-        'x': dragged.left,
-        'y': dragged.top,
-        'width': dragged.width,
-        'height': dragged.height,
-      });
-    },
-    isOwner(account) {
-      this.$store.state.accounts[account] || false;
-    },
     loadAds() {
       this.$store.commit('clearAds');
       this.contract.getAdsLength.call(function(err, res) {
@@ -143,12 +89,6 @@ export default {
         }
       }.bind(this);
       xhr.send();
-    },
-    gridStyle(config) {
-      if (!config) return;
-      return {
-        'background-image': 'url(' + config.image + ')',
-      }
     },
   },
   created() {
@@ -194,11 +134,9 @@ export default {
     }.bind(this))
   },
   components: {
-    'vue-draggable-resizable': VueDraggableResizable,
-    Ad,
-    Buy,
     Publish,
     Offline,
+    AdGrid,
   },
 }
 </script>
